♻️ sdk 환경 제거
