+++ nestia.config.ts
@@ -2,7 +2,6 @@ import type nestia from "@nestia/sdk";
 
 const NESTIA_CONFIG: nestia.INestiaConfig = {
     input: "src/controllers",
-    output: "src/api",
     json: false,
     primitive: false,
     simulate: true,
