+++ package.json
@@ -7,22 +7,18 @@
         "check": "ts-patch check",
         "prepare": "ts-patch install",
         "---------------------------------------------": "",
-        "build": "rimraf build && tsc -p ./tsconfig.build.json",
-        "build:sdk": "rimraf packages/sdk/lib && npm run nestia && tsc -p ./tsconfig.sdk.json",
         "eslint": "eslint src",
         "eslint:fix": "eslint src --fix",
         "prettier": "prettier . -c",
         "prettier:fix": "prettier . -w",
-        "dev": "npm run build && NODE_ENV=development node build/main",
         "----------------------------------------------": "",
-        "start": "NODE_ENV=production node build/main",
-        "-----------------------------------------------": "",
+        "start": "npm run build && NODE_ENV=development node build/main",
+        "build": "rimraf build && tsc -p ./tsconfig.build.json",
         "test": "rimraf tbuild && tsc -p ./tsconfig.test.json && NODE_ENV=test node tbuild/test",
-        "------------------------------------------------": "",
+        "-----------------------------------------------": "",
         "swagger": "NODE_ENV=development nestia swagger",
-        "sdk": "rimraf ./src/api/functional && NODE_ENV=development nestia sdk && find ./src/api -type f -exec sed -i '' 's/\\r$//' {} \\;",
-        "nestia": "npm run swagger ; npm run sdk",
-        "package:publish": "npm run build:api && cd ./packages/api && npm publish",
+        "swagger-ui":"node ./packages/swagger/index.mjs",
+        "doc":"npm run swagger && npm run swagger-ui",
         "-------------------------------------------------": "",
         "db:sch": "ts-node prisma",
         "db:update": "rimraf db && npm run db:sch && prisma generate && find ./db/runtime -type f -exec sed -i '' 's/\\r$//' {} \\;",
