--- packages/sdk/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@project/sdk",
-  "version": "0.0.1",
-  "description": "SDK for REST API",
-  "main": "lib/index.js",
-  "typings": "lib/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/industriously/nestia-template"
-  },
-  "author": "industriously",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/industriously/nestia-template/issues"
-  },
-  "homepage": "https://github.com/industriously/nestia-template#readme",
-  "dependencies": {
-    "@nestia/fetcher": "^1.4.0",
-    "typia": "^4.1.4"
-  }
-}
