--- src/api/HttpError.ts
@@ -1,1 +1,0 @@
-export { HttpError } from "@nestia/fetcher";
