--- src/api/IConnection.ts
@@ -1,1 +1,0 @@
-export type { IConnection } from "@nestia/fetcher";
