--- src/api/functional/auth/index.ts
@@ -1,7 +1,0 @@
-/**
- * @packageDocumentation
- * @module api.functional.auth
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
- */
-//================================================================
-export * as oauth from "./oauth";
