--- src/api/functional/auth/oauth/github/index.ts
@@ -1,61 +1,0 @@
-/**
- * @packageDocumentation
- * @module api.functional.auth.oauth.github
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
- */
-//================================================================
-import { Fetcher } from "@nestia/fetcher";
-import type { IConnection } from "@nestia/fetcher";
-import typia from "typia";
-
-/**
- * 깃허브 인증 로그인 페이지 주소 요청
- * 
- * @summary 깃허브 인증 페이지 주소 요청
- * @tag authentication
- * @return 깃허브 인증 페이지 주소
- * 
- * @controller AuthController.getGithubLoginUrl()
- * @path GET /auth/oauth/github
- * @nestia Generated by Nestia - https://github.com/samchon/nestia
- */
-export async function getGithubLoginUrl(
-    connection: IConnection,
-): Promise<getGithubLoginUrl.Output> {
-    return !!connection.simulate
-        ? getGithubLoginUrl.simulate(
-              connection,
-          )
-        : Fetcher.fetch(
-              connection,
-              getGithubLoginUrl.ENCRYPTED,
-              getGithubLoginUrl.METHOD,
-              getGithubLoginUrl.path(),
-          );
-}
-export namespace getGithubLoginUrl {
-    export type Output = string;
-
-    export const METHOD = "GET" as const;
-    export const PATH: string = "/auth/oauth/github";
-    export const ENCRYPTED: Fetcher.IEncrypted = {
-        request: false,
-        response: false,
-    };
-
-    export const path = (): string => {
-        return `/auth/oauth/github`;
-    }
-    export const random = (g?: Partial<typia.IRandomGenerator>): Output =>
-        typia.random<Output>(g);
-    export const simulate = async (
-        connection: IConnection,
-    ): Promise<Output> => {
-        return random(
-            typeof connection.simulate === 'object' &&
-                connection.simulate !== null
-                ? connection.simulate
-                : undefined
-        );
-    }
-}
