--- src/api/functional/auth/oauth/index.ts
@@ -1,8 +1,0 @@
-/**
- * @packageDocumentation
- * @module api.functional.auth.oauth
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
- */
-//================================================================
-export * as kakao from "./kakao";
-export * as github from "./github";
