--- src/api/functional/index.ts
@@ -1,8 +1,0 @@
-/**
- * @packageDocumentation
- * @module api.functional
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
- */
-//================================================================
-export * as auth from "./auth";
-export * as users from "./users";
