--- src/api/functional/users/index.ts
@@ -1,7 +1,0 @@
-/**
- * @packageDocumentation
- * @module api.functional.users
- * @nestia Generated by Nestia - https://github.com/samchon/nestia 
- */
-//================================================================
-export * as normals from "./normals";
