--- src/api/index.ts
@@ -1,5 +1,0 @@
-import * as api from "./module";
-
-export * from "./module";
-
-export default api;
