--- src/api/module.ts
@@ -1,14 +1,0 @@
-export type * from "./IConnection";
-export type * from "./Primitive";
-export * from "./HttpError";
-
-export * from "./functional";
-
-export type * from "./types/ErrorCode";
-export type * from "./types/Mutable";
-export type * from "./structures/IAuthentication";
-export type * from "./structures/IToken";
-export type * from "./structures/common/IAddress";
-export type * from "./structures/common/pagination";
-export type * from "./structures/user/INornal";
-export type * from "./structures/user/IUser";
