--- src/api/structures/IAuthentication.ts
@@ -1,3 +1,0 @@
-export namespace IAuthentication {
-    export type OuathType = "kakao" | "github";
-}
