+++ src/app/authentication/module.ts
@@ -1,0 +1,8 @@
+import { IAuthenticationService } from "./interface";
+import { Service } from "./service";
+
+export const Authentication: Authentication = { Service };
+
+export interface Authentication {
+    readonly Service: IAuthenticationService;
+}
