+++ src/app/user/module.ts
@@ -1,0 +1,9 @@
+import { IUserService } from "./interface";
+
+export const User: User = {
+    Service: {},
+};
+
+export interface User {
+    readonly Service: IUserService;
+}
