+++ src/app/user/normal/index.ts
@@ -1,0 +1,2 @@
+export * from "./module";
+export type * from "./interface";
