+++ src/app/user/normal/module.ts
@@ -1,0 +1,8 @@
+import { INormalService } from "./interface";
+import { Service } from "./service";
+
+export const Normal: Normal = { Service };
+
+export interface Normal {
+    readonly Service: INormalService;
+}
