--- src/providers/authentication/index.ts
@@ -1,11 +1,0 @@
-import { IAuthentication } from "@APP/api/structures/IAuthentication";
-
-import { Service } from "./service";
-
-export const AuthenticationService: AuthenticationService = Service;
-
-export interface AuthenticationService {
-    readonly getLoginUrl: (
-        oauth_type: IAuthentication.OuathType,
-    ) => Promise<string>;
-}
