--- tsconfig.sdk.json
@@ -1,12 +1,0 @@
-{
-    "extends": "./tsconfig.json",
-    "compilerOptions": {
-        "removeComments": false,
-        "sourceMap": false,
-        "target": "ES6",
-        "declaration": true,
-        "outDir": "./packages/sdk/lib",
-        "lib": ["DOM", "ES2015"]
-    },
-    "include": ["./src/api"]
-}
